requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.5.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
pydantic[email]==1.10.12
httpx==0.25.2
python-dotenv==1.0.0